% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_sim_data.R
\name{accum_sim_data}
\alias{accum_sim_data}
\title{Accumulate simulated data}
\usage{
accum_sim_data(data_lists)
}
\arguments{
\item{data_lists}{list of lists of dataframes (model output) from several simulation runs.}
}
\value{
a new list of lists, with some simulation runs merged.
}
\description{
Goes through a simulation set (list of lists of dataframes) and joins together experiments
with the same parameterization, adding their simulation ids in the sim_info dataframe of each.
}
\author{
Marina Papadopoulou
}
