% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{agent_config}
\alias{agent_config}
\title{Get agent type config}
\usage{
agent_config(config_list, agent_type)
}
\arguments{
\item{config_list}{a list, the config file in a named list format}

\item{agent_type}{a string, the type of agent as named in the config file, e.g. 'Pigeon'}
}
\value{
list of all agent type specific parameters
}
\description{
Extracts the config file of a specific agent type
}
\author{
Marina Papadopoulou
}
