% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_configs.R
\name{config_change}
\alias{config_change}
\title{Change a config file}
\usage{
config_change(config_temp, df_param, exp_name, bunch_id)
}
\arguments{
\item{config_temp}{a list of lists, the full config file of a simulation}

\item{df_param}{a named vector of all values to be replaced, (a line
of the output dataframe of the create_params function}

\item{exp_name}{the name with which to export the changed config}
}
\value{
none, exports a json file with the changed config
}
\description{
Replaces the existing values of the parameters in the config
with the ones from the input vector
}
\author{
Marina Papadopoulou (email = "marina.vp@outlook.com", ORCID = "0000-0002-6478-8365")
}
