% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_configs.R
\name{config_change_state}
\alias{config_change_state}
\title{Change the state level of a Prey-type agent}
\usage{
config_change_state(skchange, new_params)
}
\arguments{
\item{skchange}{a list of lists, a state from the config file of a prey-type agent}

\item{new_params}{a named vector of all values to be replaced, (a line
of the output dataframe of the create_params function}
}
\value{
the input list with the changed parameters
}
\description{
Replaces the existing values of the parameters in the configuration of a state
with the ones from the input vector
}
\author{
Marina Papadopoulou (email = "marina.vp@outlook.com", ORCID = "0000-0002-6478-8365")
}
