% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_class.R
\name{construct_ParamSets}
\alias{construct_ParamSets}
\title{Construct param set}
\usage{
construct_ParamSets(
  N = NA_integer_,
  Tmax = NA_integer_,
  aer_cs_sd = NA_integer_,
  fl_aer_cs = NA_integer_,
  fl_aer_w = NA_integer_,
  coh_speed_w = NA_integer_,
  coh_speed_topo = NA_integer_,
  coh_speed_ffov = NA_integer_,
  coh_speed_mindist = NA_integer_,
  coh_speed_maxdist = NA_integer_,
  roost_w = NA_integer_,
  fl_tr = NA_integer_,
  coh_turn_topo = NA_integer_,
  coh_turn_w = NA_integer_,
  sep_w = NA_integer_,
  sep_mindist = NA_integer_,
  sep_topo = NA_integer_,
  ali_topo = NA_integer_,
  ali_w = NA_integer_,
  output_folder = NA_character_,
  wig_w = NA_integer_,
  roost_s_freq = NA_integer_,
  turn_dur = NA_integer_,
  deg2home = NA_integer_,
  dist2home = NA_integer_,
  fov = NA_integer_,
  turn_chars = NA_integer_,
  copy_fov = NA_integer_,
  copy_topo = NA_integer_,
  Npred = NA_integer_
)
}
\description{
Construct a new ParamSets object
}
