% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param_set.R
\name{create_param_set}
\alias{create_param_set}
\title{Create parameter set}
\usage{
create_param_set(ps)
}
\arguments{
\item{ps}{Parameters set class object}
}
\value{
a column-named dataframe with each conbination of the input parameters (as rows)
}
\description{
Creates a data frame with all the combinations of each parameter vector
}
\author{
Marina Papadopoulou
}
