% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sim_data.R
\name{discard_init_timesteps}
\alias{discard_init_timesteps}
\title{Discard initial timesteps}
\usage{
discard_init_timesteps(data_list, cut_time = 1)
}
\arguments{
\item{data_list}{list of base model data, export of import_base_data}

\item{cut_time}{the time point in seconds from which onwards to keep in the dataframes}
}
\value{
the input dataframe cut after the init cut point
}
\description{
Discards timesteps from the beginning of the simulation until a cut point from all the dataframes exported by the model.
}
\author{
Marina Papadopoulou
}
