% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_parameters}
\alias{get_parameters}
\title{Get simulation parameters}
\usage{
get_parameters(config, verbose = FALSE)
}
\arguments{
\item{config}{the config file of a simulation}
}
\value{
a named list with all parameters of interest
}
\description{
Extract parameters of interest from the config file
}
\author{
Marina Papadopoulou
}
