% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params_vector}
\alias{get_params_vector}
\title{Get parameters vector}
\usage{
get_params_vector(params_list, par_order)
}
\arguments{
\item{params_list}{named list with all parameters of interest, result of get_parameters}

\item{par_order}{named list with all parameters of interest}
}
\value{
a vector with the parameter values in the same order as the par_order
}
\description{
Export ordered vector from list of parameters based on input of parameter names
}
\author{
Marina Papadopoulou
}
