% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{get_state}
\alias{get_state}
\title{Get state config}
\usage{
get_state(agent_config, state_n)
}
\arguments{
\item{agent_config}{a list, the config of a specific agent-type in a named list format}

\item{state_n}{the number of a state}
}
\value{
list of all parameters (the config) of a specific state of an agent type
}
\description{
Get a specific state of an agent
}
\author{
Marina Papadopoulou
}
