% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sim_data.R
\name{import_base_data}
\alias{import_base_data}
\title{Import simulated data}
\usage{
import_base_data(
  folder_path,
  types = c("self", "neighb", "group"),
  config_name = "composed_config.json"
)
}
\arguments{
\item{folder_path}{path to data folder}

\item{types}{which data to load, from: timeseries, neighbours, flocks, roosting, curvature, transition matrix and forces.}
}
\value{
list of all imported dataframes
}
\description{
Imports all the data produced by a model simulation.
}
\author{
Marina Papadopoulou
}
