% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sim_data.R
\name{import_multi_data}
\alias{import_multi_data}
\title{Import multiple simulated datasets}
\usage{
import_multi_data(head_folder_path, types = c("self"))
}
\arguments{
\item{head_folder_path}{path to result folder that contains all the single simulation folder results.}

\item{types}{which data to load, from: timeseries, neighbours, flocks, roosting, curvature, transition matrix and forces.}
}
\value{
list of all simulation lists
}
\description{
Imports all the data saved from multiple model simulations. Calls import_base_data for each folder
}
\author{
Marina Papadopoulou
}
