% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{load_config}
\alias{load_config}
\title{Load config.json file}
\usage{
load_config(path2config)
}
\arguments{
\item{path2config}{path to json config file}
}
\value{
list of all model parameters
}
\description{
Loads config file from json format used to generate simulated data, turning it into a list
}
\author{
Marina Papadopoulou
}
