% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_class.R
\docType{class}
\name{ParamSets}
\alias{ParamSets}
\title{Initialize the class with all parameters that can be changed.}
\description{
Class for the parameters of the computational model
}
\keyword{class}
