% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_configs.R
\name{prey_config_change}
\alias{prey_config_change}
\title{Change the config of a Prey-type agent}
\usage{
prey_config_change(conf_prey, new_params)
}
\arguments{
\item{conf_prey}{a list of lists, the config file of a prey-type agent}

\item{new_params}{a named vector of all values to be replaced, (a line
of the output dataframe of the create_params function}
}
\value{
the config file (list of lists) with the changed parameters
}
\description{
Replaces the existing values of the parameters in the config
with the ones from the input vector
}
\author{
Marina Papadopoulou (email = "marina.vp@outlook.com", ORCID = "0000-0002-6478-8365")
}
