% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations}
\usage{
run_simulations(dirs_list, reps)
}
\arguments{
\item{reps}{integer; How many times to repeat each simulation with same parameters.}

\item{dir_list}{A list of directories exported by the set_up_dirs function.}
}
\value{
nothing, it runs the simulations and writes a log file.
}
\description{
Runs simulations through the shell.
}
\author{
Marina Papadopoulou
}
