% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_dirs.R
\name{set_up_directories}
\alias{set_up_directories}
\title{Set up directories for simulations}
\usage{
set_up_directories(
  sim_name,
  config_name,
  main_path,
  model_exe,
  exp_conf_name_id = ""
)
}
\arguments{
\item{sim_name}{the name of the simulation experiment (to be used as unique id of folder).}

\item{config_name}{the name of the configuration file.}

\item{main_path}{the main directory to use for all files.}

\item{model_exe}{the name of the model executable.}

\item{exp_conf_name_id}{optional, if a special id is needed in the name of the generated configs.}
}
\value{
a list of strings, with all paths to be used.
}
\description{
Collects the paths to the configs and model executable, and sets up the
working directory using the input simulation name to store the generated configs and
}
\author{
Marina Papadopoulou
}
