% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_sims.R
\name{set_up_simulations}
\alias{set_up_simulations}
\title{Set up simulations}
\usage{
set_up_simulations(param_set_class, dirs_list)
}
\arguments{
\item{param_set_class}{a parameter class with values to vary.}

\item{dirs_list}{a list of paths as exported by the set_up_dirs function.}
}
\value{
nothing, creates as many config files as parameter combinations.
}
\description{
Creates configuration files to be used for simulation runs.
}
\author{
Marina Papadopoulou
}
