% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{sim_config}
\alias{sim_config}
\title{Get the simulation config}
\usage{
sim_config(config_list)
}
\arguments{
\item{config_list}{a list, the config file in a named list format}
}
\value{
list of all simulation parameters
}
\description{
Extracts the simulation part of the config file.
}
\author{
Marina Papadopoulou
}
