% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{state_actions}
\alias{state_actions}
\title{Actions of a state}
\usage{
state_actions(agent_config, state_n)
}
\arguments{
\item{agent_config}{a list, the config of a specific agent-type in a named list format}

\item{state_n}{the number of a state}
}
\value{
a vector with the names of all actions included in the state
}
\description{
Get the actions included in a specific state
}
\author{
Marina Papadopoulou
}
