% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{state_actions_names}
\alias{state_actions_names}
\title{Actions of a state}
\usage{
state_actions_names(state_config)
}
\arguments{
\item{state_config}{a list, the config of a specific state of a specific agent-type in a named list format}
}
\value{
a vector with the names of all actions included in the state
}
\description{
Get the actions included in a specific state
}
\author{
Marina Papadopoulou
}
