% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{state_descr}
\alias{state_descr}
\title{State description}
\usage{
state_descr(agent_config, state_n)
}
\arguments{
\item{agent_config}{a list, the config of a specific agent-type in a named list format}

\item{state_n}{the number of a state}
}
\value{
a string with the description of the state
}
\description{
Get the description of all actions included in a specific state
}
\author{
Marina Papadopoulou
}
