% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{states_composition}
\alias{states_composition}
\title{State composition}
\usage{
states_composition(agent_config)
}
\arguments{
\item{agent_config}{a list, the config of a specific agent-type in a named list format}
}
\value{
list of all states with their actions composition
}
\description{
Get the composition of actions of a specific state of an agent type
}
\author{
Marina Papadopoulou
}
