% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{states_size}
\alias{states_size}
\title{Number of states}
\usage{
states_size(agent_config)

states_size(agent_config)
}
\arguments{
\item{agent_config}{a list, the config of a specific agent-type in a named list format}

\item{config_list}{a list, the config file in a named list format}

\item{agent_type}{a string, the type of agent as named in the config file, e.g. 'Pigeon'}
}
\value{
numeric, the number of potential states of an agent

numeric, the number of potential states of the agent
}
\description{
Get the number of states of a specific agent-type.

Get the number of states of a given agent type config.
}
\author{
Marina Papadopoulou
}
